refactor: Improve code modularity and fix e2e test reliability
Split large files into focused modules for better maintainability:
- Extract AppShell agent/terminal operations into useAgentOperations
  and useTerminalOperations hooks
- Split agent_messaging.go into agent_delivery, agent_output, and
  agent_plan_mode modules
- Split workspace_service.go into workspace_agent_ops and
  workspace_watch modules
- Extract editor draft persistence into draftPersistence.ts
- Extract message rendering utilities into messageUtils.ts
- Remove unused email/sender.go and sanitize packages

Fix flaky e2e tests and improve plan mode test prompts:
- Increase waitForControlBanner timeout to 60s under parallel load
- Use dummy plan content so plan execution finishes quickly
- Extract shared PLAN_MODE_PROMPT constant and planModePrompt()
  function into helpers.ts with documentation
- Remove e2e tests from CI (no coding agent access in GitHub Actions)
