+++ .github/workflows/ci.yaml
@@ -72,21 +72,3 @@ jobs:
 
       - name: Test
         run: task test
-
-      - name: Install Playwright browsers
-        if: github.event_name == 'push'
-        run: cd frontend && bunx playwright install --with-deps chromium
-
-      - name: Run e2e tests
-        if: github.event_name == 'push'
-        run: task test-e2e
-
-      - name: Upload Playwright report
-        if: github.event_name == 'push' && failure()
-        uses: actions/upload-artifact@v4
-        with:
-          name: playwright-report
-          path: |
-            frontend/playwright-report/
-            frontend/test-results/
-          retention-days: 7
