+++ internal/hub/service/workspace_service.go
@@ -5,7 +5,6 @@ import (
 	"database/sql"
 	"fmt"
 	"log/slog"
-	"sync"
 
 	"connectrpc.com/connect"
 	"google.golang.org/protobuf/encoding/protojson"
@@ -15,7 +14,6 @@ import (
 	"github.com/leapmux/leapmux/internal/hub/auth"
 	"github.com/leapmux/leapmux/internal/hub/generated/db"
 	"github.com/leapmux/leapmux/internal/hub/id"
-	"github.com/leapmux/leapmux/internal/hub/layout"
 	"github.com/leapmux/leapmux/internal/hub/lexorank"
 	"github.com/leapmux/leapmux/internal/hub/terminalmgr"
 	"github.com/leapmux/leapmux/internal/hub/validate"
@@ -629,496 +627,6 @@ func (s *WorkspaceService) GetLayout(
 		Layout:     &layoutNode,
 		ActiveTabs: activeTabs,
 	}), nil
-}
-
-func (s *WorkspaceService) SaveLayout(
-	ctx context.Context,
-	req *connect.Request[leapmuxv1.SaveLayoutRequest],
-) (*connect.Response[leapmuxv1.SaveLayoutResponse], error) {
-	user, err := auth.MustGetUser(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	workspaceID := req.Msg.GetWorkspaceId()
-	ws, err := s.getVisibleWorkspace(ctx, user, req.Msg.GetOrgId(), workspaceID)
-	if err != nil {
-		return nil, err
-	}
-
-	layoutNode := req.Msg.GetLayout()
-	if layoutNode == nil {
-		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("layout is required"))
-	}
-
-	// Validate the layout structure.
-	if err := layout.Validate(layoutNode); err != nil {
-		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("invalid layout: %w", err))
-	}
-
-	// Reject unoptimized layouts — the frontend must send canonical form.
-	if !layout.IsOptimized(layoutNode) {
-		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("layout is not in canonical (optimized) form"))
-	}
-
-	// Serialize to JSON.
-	jsonBytes, err := protojson.Marshal(layoutNode)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("marshal layout: %w", err))
-	}
-
-	if err := s.queries.UpsertWorkspaceLayout(ctx, db.UpsertWorkspaceLayoutParams{
-		WorkspaceID: workspaceID,
-		LayoutJson:  string(jsonBytes),
-	}); err != nil {
-		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("save layout: %w", err))
-	}
-
-	// Save per-tile active tabs.
-	if err := s.queries.DeleteTileActiveTabs(ctx, workspaceID); err != nil {
-		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("clear tile active tabs: %w", err))
-	}
-
-	for _, at := range req.Msg.GetActiveTabs() {
-		if err := s.queries.UpsertTileActiveTab(ctx, db.UpsertTileActiveTabParams{
-			WorkspaceID: workspaceID,
-			TileID:      at.TileId,
-			TabType:     at.TabType,
-			TabID:       at.TabId,
-		}); err != nil {
-			return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("save tile active tab: %w", err))
-		}
-	}
-
-	// Reconcile workspace tabs if the request includes them.
-	if tabs := req.Msg.GetTabs(); len(tabs) > 0 {
-		if err := s.queries.DeleteWorkspaceTabsByWorkspace(ctx, workspaceID); err != nil {
-			return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("clear workspace tabs: %w", err))
-		}
-		for _, tab := range tabs {
-			if err := s.queries.UpsertWorkspaceTab(ctx, db.UpsertWorkspaceTabParams{
-				WorkspaceID: workspaceID,
-				TabType:     tab.GetTabType(),
-				TabID:       tab.GetTabId(),
-				Position:    tab.GetPosition(),
-				TileID:      tab.GetTileId(),
-			}); err != nil {
-				return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("save workspace tab: %w", err))
-			}
-		}
-	}
-
-	// Clean up agents and terminals whose tiles no longer exist in the layout.
-	s.cleanupRemovedTiles(ctx, ws, layoutNode)
-
-	return connect.NewResponse(&leapmuxv1.SaveLayoutResponse{}), nil
-}
-
-// cleanupRemovedTiles finds workspace_tabs whose tile_id is not in the new
-// layout and closes the associated agents/terminals.
-func (s *WorkspaceService) cleanupRemovedTiles(ctx context.Context, ws *db.Workspace, layoutNode *leapmuxv1.LayoutNode) {
-	tileIDs := layout.CollectLeafIDs(layoutNode)
-
-	tabs, err := s.queries.ListWorkspaceTabsByWorkspace(ctx, ws.ID)
-	if err != nil {
-		slog.Error("failed to list workspace tabs for tile cleanup", "workspace_id", ws.ID, "error", err)
-		return
-	}
-
-	for _, tab := range tabs {
-		if tab.TileID == "" {
-			continue // Tab not assigned to a tile; skip.
-		}
-		if _, ok := tileIDs[tab.TileID]; ok {
-			continue // Tile still exists in the layout.
-		}
-
-		// This tab's tile has been removed — clean up.
-		switch tab.TabType {
-		case leapmuxv1.TabType_TAB_TYPE_AGENT:
-			agent, agentErr := s.queries.GetAgentByID(ctx, tab.TabID)
-			var conn *workermgr.Conn
-			if agentErr == nil {
-				conn = s.workerMgr.Get(agent.WorkerID)
-			} else {
-				slog.Warn("failed to get agent for tile cleanup", "agent_id", tab.TabID, "error", agentErr)
-			}
-			s.closeAgentForTileCleanup(ctx, ws, conn, tab.TabID)
-		case leapmuxv1.TabType_TAB_TYPE_TERMINAL:
-			var conn *workermgr.Conn
-			if s.termSvc != nil {
-				if workerID := s.termSvc.GetTerminalWorkerID(tab.TabID); workerID != "" {
-					conn = s.workerMgr.Get(workerID)
-				}
-			}
-			s.closeTerminalForTileCleanup(ctx, ws, conn, tab.TabID)
-		}
-
-		// Remove the workspace_tab entry.
-		if err := s.queries.DeleteWorkspaceTab(ctx, db.DeleteWorkspaceTabParams{
-			WorkspaceID: ws.ID,
-			TabType:     tab.TabType,
-			TabID:       tab.TabID,
-		}); err != nil {
-			slog.Warn("failed to delete workspace tab during tile cleanup",
-				"workspace_id", ws.ID, "tab_type", tab.TabType, "tab_id", tab.TabID, "error", err)
-		}
-	}
-}
-
-// closeAgentForTileCleanup stops an agent on the worker, marks it closed in
-// the DB, and broadcasts a status change to watchers.
-func (s *WorkspaceService) closeAgentForTileCleanup(ctx context.Context, ws *db.Workspace, conn *workermgr.Conn, agentID string) {
-	if conn != nil {
-		if err := conn.Send(&leapmuxv1.ConnectResponse{
-			Payload: &leapmuxv1.ConnectResponse_AgentStop{
-				AgentStop: &leapmuxv1.AgentStopRequest{
-					WorkspaceId: ws.ID,
-					AgentId:     agentID,
-				},
-			},
-		}); err != nil {
-			slog.Warn("failed to send agent stop during tile cleanup", "agent_id", agentID, "error", err)
-		}
-	}
-
-	if err := s.queries.CloseAgent(ctx, agentID); err != nil {
-		slog.Warn("failed to close agent during tile cleanup", "agent_id", agentID, "error", err)
-	}
-
-	if err := s.queries.DeleteControlRequestsByAgentID(ctx, agentID); err != nil {
-		slog.Warn("failed to delete control requests during tile cleanup", "agent_id", agentID, "error", err)
-	}
-
-	// Read agent for status broadcast (best-effort).
-	agent, err := s.queries.GetAgentByID(ctx, agentID)
-	if err == nil {
-		tileSc := AgentStatusChange(&agent, conn != nil)
-		if s.agentSvc != nil {
-			tileSc.GitStatus = s.agentSvc.GetGitStatus(agentID)
-		}
-		s.agentMgr.Broadcast(agentID, &leapmuxv1.AgentEvent{
-			AgentId: agentID,
-			Event: &leapmuxv1.AgentEvent_StatusChange{
-				StatusChange: tileSc,
-			},
-		})
-	}
-
-	// Best-effort worktree cleanup (no user confirmation during tile cleanup).
-	s.worktreeHelper.UnregisterTabBestEffort(ctx, leapmuxv1.TabType_TAB_TYPE_AGENT, agentID)
-}
-
-// closeTerminalForTileCleanup stops a terminal on the worker and broadcasts
-// a closed event to watchers.
-func (s *WorkspaceService) closeTerminalForTileCleanup(ctx context.Context, _ *db.Workspace, conn *workermgr.Conn, terminalID string) {
-	if s.termSvc != nil {
-		s.termSvc.CloseTerminalInternal(ctx, conn, terminalID)
-	}
-}
-
-// watchEventsMerged is the internal type sent through the merged channel
-// by per-watcher goroutines in WatchEvents.
-type watchEventsMerged struct {
-	agentEvent    *leapmuxv1.AgentEvent
-	terminalEvent *leapmuxv1.TerminalEvent
-}
-
-// watchEventsCore contains the core WatchEvents logic decoupled from transport.
-// send is called for each WatchEventsResponse; return an error to stop the stream.
-func (s *WorkspaceService) watchEventsCore(
-	ctx context.Context,
-	user *auth.UserInfo,
-	req *leapmuxv1.WatchEventsRequest,
-	send func(*leapmuxv1.WatchEventsResponse) error,
-) error {
-	ws, err := s.getVisibleWorkspace(ctx, user, req.GetOrgId(), req.GetWorkspaceId())
-	if err != nil {
-		return err
-	}
-
-	// Collect entries, limiting total to 32.
-	const maxEntries = 32
-	agentEntries := req.GetAgents()
-	terminalEntries := req.GetTerminals()
-	total := len(agentEntries) + len(terminalEntries)
-	if total > maxEntries {
-		// Truncate terminal entries first, then agent entries.
-		if len(terminalEntries) > maxEntries-len(agentEntries) {
-			if len(agentEntries) >= maxEntries {
-				agentEntries = agentEntries[:maxEntries]
-				terminalEntries = nil
-			} else {
-				terminalEntries = terminalEntries[:maxEntries-len(agentEntries)]
-			}
-		}
-	}
-
-	// Deduplicate agent IDs.
-	seenAgents := make(map[string]struct{}, len(agentEntries))
-	var dedupAgents []*leapmuxv1.WatchAgentEntry
-	for _, e := range agentEntries {
-		aid := e.GetAgentId()
-		if _, ok := seenAgents[aid]; ok {
-			continue
-		}
-		seenAgents[aid] = struct{}{}
-		dedupAgents = append(dedupAgents, e)
-	}
-	agentEntries = dedupAgents
-
-	// Deduplicate terminal IDs.
-	seenTerminals := make(map[string]struct{}, len(terminalEntries))
-	var dedupTerminals []*leapmuxv1.WatchTerminalEntry
-	for _, e := range terminalEntries {
-		tid := e.GetTerminalId()
-		if _, ok := seenTerminals[tid]; ok {
-			continue
-		}
-		seenTerminals[tid] = struct{}{}
-		dedupTerminals = append(dedupTerminals, e)
-	}
-	terminalEntries = dedupTerminals
-
-	// --- Register watchers before replay to capture live events ---
-	type agentWatcher struct {
-		agentID string
-		watcher *agentmgr.Watcher
-	}
-	type termWatcher struct {
-		terminalID string
-		watcher    *terminalmgr.Watcher
-	}
-
-	var agentWatchers []agentWatcher
-	for _, entry := range agentEntries {
-		agentID := entry.GetAgentId()
-		w := s.agentMgr.Watch(agentID)
-		agentWatchers = append(agentWatchers, agentWatcher{agentID: agentID, watcher: w})
-	}
-	defer func() {
-		for _, aw := range agentWatchers {
-			s.agentMgr.Unwatch(aw.agentID, aw.watcher)
-		}
-	}()
-
-	var termWatchers []termWatcher
-	for _, entry := range terminalEntries {
-		terminalID := entry.GetTerminalId()
-		w := s.termMgr.Watch(terminalID)
-		termWatchers = append(termWatchers, termWatcher{terminalID: terminalID, watcher: w})
-	}
-	defer func() {
-		for _, tw := range termWatchers {
-			s.termMgr.Unwatch(tw.terminalID, tw.watcher)
-		}
-	}()
-
-	// Track replayed control request IDs per agent so we can deduplicate
-	// against live events that arrived in the watcher channel during replay.
-	replayedCRs := make(map[string]map[string]struct{}) // agentID -> set of requestIDs
-
-	// --- Replay historical data and send snapshots for agents ---
-	for _, entry := range agentEntries {
-		agentID := entry.GetAgentId()
-		agent, err := s.queries.GetAgentByID(ctx, agentID)
-		if err != nil {
-			if err == sql.ErrNoRows {
-				slog.Debug("watchEventsCore: agent not found, skipping", "agent_id", agentID)
-				continue
-			}
-			return fmt.Errorf("get agent %s: %w", agentID, err)
-		}
-		if agent.WorkspaceID != ws.ID {
-			slog.Debug("watchEventsCore: agent not in workspace, skipping", "agent_id", agentID, "workspace_id", ws.ID)
-			continue
-		}
-
-		// Send historical messages if requested (afterSeq >= 0).
-		afterSeq := entry.GetAfterSeq()
-		if afterSeq >= 0 {
-			msgs, err := s.queries.ListMessagesByAgentID(ctx, db.ListMessagesByAgentIDParams{
-				AgentID: agentID,
-				Seq:     afterSeq,
-				Limit:   50,
-			})
-			if err != nil {
-				return fmt.Errorf("list messages for agent %s: %w", agentID, err)
-			}
-			for i := range msgs {
-				if err := send(&leapmuxv1.WatchEventsResponse{
-					Event: &leapmuxv1.WatchEventsResponse_AgentEvent{
-						AgentEvent: &leapmuxv1.AgentEvent{
-							AgentId: agentID,
-							Event: &leapmuxv1.AgentEvent_AgentMessage{
-								AgentMessage: MessageToProto(&msgs[i]),
-							},
-						},
-					},
-				}); err != nil {
-					return err
-				}
-			}
-		}
-
-		// Send current status snapshot (check worker online per-agent).
-		workerOnline := s.workerMgr.IsOnline(agent.WorkerID)
-		catchupSc := AgentStatusChange(&agent, workerOnline)
-		if s.agentSvc != nil {
-			catchupSc.GitStatus = s.agentSvc.GetGitStatus(agentID)
-		}
-		if err := send(&leapmuxv1.WatchEventsResponse{
-			Event: &leapmuxv1.WatchEventsResponse_AgentEvent{
-				AgentEvent: &leapmuxv1.AgentEvent{
-					AgentId: agentID,
-					Event: &leapmuxv1.AgentEvent_StatusChange{
-						StatusChange: catchupSc,
-					},
-				},
-			},
-		}); err != nil {
-			return err
-		}
-
-		// Replay pending control requests and track their IDs.
-		pendingCRs, err := s.queries.ListControlRequestsByAgentID(ctx, agentID)
-		if err != nil {
-			slog.Warn("watchEventsCore: list pending control requests", "agent_id", agentID, "error", err)
-		} else {
-			for _, cr := range pendingCRs {
-				if replayedCRs[agentID] == nil {
-					replayedCRs[agentID] = make(map[string]struct{})
-				}
-				replayedCRs[agentID][cr.RequestID] = struct{}{}
-
-				if err := send(&leapmuxv1.WatchEventsResponse{
-					Event: &leapmuxv1.WatchEventsResponse_AgentEvent{
-						AgentEvent: &leapmuxv1.AgentEvent{
-							AgentId: agentID,
-							Event: &leapmuxv1.AgentEvent_ControlRequest{
-								ControlRequest: &leapmuxv1.AgentControlRequest{
-									AgentId:   cr.AgentID,
-									RequestId: cr.RequestID,
-									Payload:   cr.Payload,
-								},
-							},
-						},
-					},
-				}); err != nil {
-					return err
-				}
-			}
-		}
-	}
-
-	// --- Drain buffered watcher events that arrived during replay ---
-	// Any controlRequest events whose requestId was already replayed are
-	// duplicates and must be skipped; all other events are forwarded.
-	for _, aw := range agentWatchers {
-		crSet := replayedCRs[aw.agentID]
-	drain:
-		for {
-			select {
-			case ev := <-aw.watcher.C():
-				if cr := ev.GetControlRequest(); cr != nil && crSet != nil {
-					if _, dup := crSet[cr.GetRequestId()]; dup {
-						continue // skip duplicate
-					}
-				}
-				if err := send(&leapmuxv1.WatchEventsResponse{
-					Event: &leapmuxv1.WatchEventsResponse_AgentEvent{
-						AgentEvent: ev,
-					},
-				}); err != nil {
-					return err
-				}
-			default:
-				break drain
-			}
-		}
-	}
-
-	// Merged channel collects events from all watchers.
-	merged := make(chan watchEventsMerged, 64)
-
-	var wg sync.WaitGroup
-
-	// Spawn goroutines that forward agent watcher events into merged.
-	for _, aw := range agentWatchers {
-		wg.Add(1)
-		go func(aw agentWatcher) {
-			defer wg.Done()
-			for {
-				select {
-				case <-ctx.Done():
-					return
-				case ev, ok := <-aw.watcher.C():
-					if !ok {
-						return
-					}
-					select {
-					case merged <- watchEventsMerged{agentEvent: ev}:
-					case <-ctx.Done():
-						return
-					}
-				}
-			}
-		}(aw)
-	}
-
-	// Spawn goroutines that forward terminal watcher events into merged.
-	for _, tw := range termWatchers {
-		wg.Add(1)
-		go func(tw termWatcher) {
-			defer wg.Done()
-			for {
-				select {
-				case <-ctx.Done():
-					return
-				case ev, ok := <-tw.watcher.C():
-					if !ok {
-						return
-					}
-					select {
-					case merged <- watchEventsMerged{terminalEvent: ev}:
-					case <-ctx.Done():
-						return
-					}
-				}
-			}
-		}(tw)
-	}
-
-	// Close merged channel once all forwarders exit.
-	go func() {
-		wg.Wait()
-		close(merged)
-	}()
-
-	// Stream loop: read from merged, wrap and send.
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case m, ok := <-merged:
-			if !ok {
-				return nil
-			}
-			var resp leapmuxv1.WatchEventsResponse
-			if m.agentEvent != nil {
-				resp.Event = &leapmuxv1.WatchEventsResponse_AgentEvent{
-					AgentEvent: m.agentEvent,
-				}
-			} else if m.terminalEvent != nil {
-				resp.Event = &leapmuxv1.WatchEventsResponse_TerminalEvent{
-					TerminalEvent: m.terminalEvent,
-				}
-			}
-			if err := send(&resp); err != nil {
-				return err
-			}
-		}
-	}
 }
 
 func workspaceToProto(w *db.Workspace) *leapmuxv1.Workspace {
